requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.10
django-cors-headers==3.10.1
djangorestframework==3.13.1
python-dotenv==0.19.2
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.3.2
numpy==1.21.5
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
uWSGI>=2.0.19.1,<2.1